requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pypdf>=4.0.0
python-docx>=1.0.0
striprtf>=0.0.26
